asm/c128: Adding asm stubs for Dot functions.
