+++ internal/asm/c128/stubs_amd64.go
@@ -62,3 +62,35 @@ func ScalInc(alpha complex128, x []complex128, n, inc uintptr)
 //  	x[i] *= alpha
 //  }
 func ScalUnitary(alpha complex128, x []complex128)
+
+// DotcUnitary is
+//  for i, v := range x {
+//  	sum += y[i] * cmplx.Conj(v)
+//  }
+//  return sum
+func DotcUnitary(x, y []complex128) (sum complex128)
+
+// DotcInc is
+//  for i := 0; i < int(n); i++ {
+//  	sum += y[iy] * cmplx.Conj(x[ix])
+//  	ix += incX
+//  	iy += incY
+//  }
+//  return sum
+func DotcInc(x, y []complex128, n, incX, incY, ix, iy uintptr) (sum complex128)
+
+// DotuUnitary is
+//  for i, v := range x {
+//  	sum += y[i] * v
+//  }
+//  return sum
+func DotuUnitary(x, y []complex128) (sum complex128)
+
+// DotuInc is
+//  for i := 0; i < int(n); i++ {
+//  	sum += y[iy] * x[ix]
+//  	ix += incX
+//  	iy += incY
+//  }
+//  return sum
+func DotuInc(x, y []complex128, n, incX, incY, ix, iy uintptr) (sum complex128)
